@@ -245,7 +245,12 @@ export function calculatePage(state: State): number {
         return 1;
     }
 
-    return Math.ceil((state.page.to + 1) / state.page.size);
+    let pageNumber = Math.ceil((state.page.to + 1) / state.page.size);
+    if (pageNumber === 0) {
+        return 1;
+    } else {
+        return pageNumber;
+    }
 }
 
 /**
