@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { VulnerabilityItem, VulnerabilitySeverity } from '../service/index';
-
+import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { ResultGridComponent } from './result-grid.component';
 import { ScanningResultService, ScanningResultDefaultService } from '../service/scanning.service';
 import { SERVICE_CONFIG, IServiceConfig } from '../service.config';
@@ -26,7 +26,8 @@ describe('ResultGridComponent (inline template)', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [
-        SharedModule
+        SharedModule,
+        NoopAnimationsModule
       ],
       declarations: [ResultGridComponent, FilterComponent],
       providers: [
