@@ -109,12 +109,9 @@ export class AuditLogComponent implements OnInit {
       );
   }
 
-  retrievePage(state: State) {
-    if (state && state.page) {
-      this.queryParam.page = Math.ceil((state.page.to + 1) / this.pageSize);
-      this.currentPage = this.queryParam.page;
-      this.retrieve();
-    }
+  retrievePage() {
+    this.queryParam.page = this.currentPage;
+    this.retrieve();
   }
 
   doSearchAuditLogs(searchUsername: string): void {
